+++ test/dynamo/test_deviceguard.py
@@ -6,7 +6,7 @@ import torch
 import torch._dynamo.test_case
 import torch._dynamo.testing
 from torch._dynamo.device_interface import CudaInterface, DeviceGuard
-from torch.testing._internal.common_cuda import TEST_CUDA, TEST_MULTIGPU
+from torch.testing._internal.common_cuda import TEST_CUDA
 
 
 class TestDeviceGuard(torch._dynamo.test_case.TestCase):
@@ -55,21 +55,6 @@ class TestCUDADeviceGuard(torch._dynamo.test_case.TestCase):
     def setUp(self):
         super().setUp()
         self.device_interface = CudaInterface
-
-    @unittest.skipIf(not TEST_MULTIGPU, "need multiple GPU")
-    def test_device_guard(self):
-        current_device = torch.cuda.current_device()
-
-        device_guard = DeviceGuard(self.device_interface, 1)
-
-        with device_guard as _:
-            self.assertEqual(torch.cuda.current_device(), 1)
-            self.assertEqual(device_guard.prev_idx, 0)
-            self.assertEqual(device_guard.idx, 1)
-
-        self.assertEqual(torch.cuda.current_device(), current_device)
-        self.assertEqual(device_guard.prev_idx, 0)
-        self.assertEqual(device_guard.idx, 1)
 
     def test_device_guard_no_index(self):
         current_device = torch.cuda.current_device()
